feat(extract): add MiniMax provider to CloudApiExtractor, share the fixed prompt, and make max_tokens configurable
Wires the new lunaris_llm::CloudProvider::MiniMax variant through the
extract-crate-local CloudProvider enum (FromStr, default_model,
api_key_env, model_env, and the private bridge From impl into
lunaris_llm::CloudProvider), so CloudApiExtractor can target MiniMax's
api.minimax.io endpoint directly instead of via the Ollama-shaped
tmp/route_shim.py used for prototyping. CloudApiExtractor is the
architecturally correct target for cloud-routed extraction: unlike
OllamaExtractor/LlmExtractor, it emits a sentinel entity on
retry-exhaustion instead of silently returning an empty extraction.

While wiring MiniMax in, found that cloud_api.rs's
extract_one_with_sentinel held its own independent, equally vague
extraction prompt (no field names) -- the same defect just fixed in
llm_extractor.rs::build_prompt, just never caught here because nothing
had live-tested this code path against a real model either. Made
build_prompt pub(crate) and reused it from cloud_api.rs instead of
maintaining a second copy.

Also adds a configurable max_tokens (CloudApiExtractorOpts.max_tokens,
default 512) after observing MiniMax return finish_reason: length /
empty content on some chunks with the previous hardcoded 512-token
generation ceiling during live A/B testing.

Fixed two adjacent clippy findings surfaced by the workspace-wide
rebuild this change triggers: an unnecessary u32->u32 cast in
lunaris-llm/src/cloud/anthropic.rs, and a cloud-api-only (missing
extractor-it) unused-import gate in
lunaris-extract/tests/extractor_contract.rs -- and a missing
timeout_ms field in that same file's ollama_extracts_real_batch,
left over from the earlier OllamaExtractorOpts timeout-field change.

Red/green TDD throughout: every new field/variant confirmed to fail
compilation before implementing. cargo test -p lunaris-extract --lib
--features cloud-api,ollama: 48/48 pass. cargo clippy --workspace
--all-targets -- -D warnings: clean.

author: Tin Dang

+++ crates/lunaris-llm/src/cloud/anthropic.rs
@@ -21,7 +21,7 @@ pub(super) fn build_request(
     // max_tokens is required for /v1/messages. The caller's GenOpts
     // value already accounts for the worst-case output the upstream
     // schema can produce.
-    let max_tokens = opts.max_tokens as u32;
+    let max_tokens = opts.max_tokens;
 
     let body = match constraint {
         SchemaConstraint::JsonSchema(schema) => serde_json::json!({
